+++ firmware/include/ai.h
