+++ firmware/src/main.cpp
@@ -1,4 +1,3 @@
-#include "user_app.h" // Include the user application header
 #include <Arduino.h>
 #include <ArduinoOTA.h>
 #include <ESPmDNS.h>
@@ -7,54 +6,22 @@
 #include <WebServer.h>
 #include <WiFi.h>
 #include <WiFiUdp.h>
-
+#include "ai.h"
 
 // wifi credentials
 const char *ssid = WIFI_SSID;
 const char *password = WIFI_PASSWORD;
 
-const int LED_PIN = 2; // the built-in led
-
 WebServer server(80);
 
-// --- Handlers ---
-// These handlers can control the 'isUserAppActive' flag to enable/disable user
-// logic
-
-void handleLedOn() {
-  isUserAppActive = true;
-  // Reset user app state if needed (optional)
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(200, "text/plain", "User App Active (Morse/Effects ON)");
-  Serial.println("User App ON");
-}
-
-void handleLedOff() {
-  isUserAppActive = false;
-  digitalWrite(LED_PIN, LOW); // Turn off immediately
-  server.sendHeader("Access-Control-Allow-Origin", "*");
-  server.send(200, "text/plain", "User App Inactive (LED OFF)");
-  Serial.println("User App OFF");
-}
-
-void handleToggle() {
-  if (isUserAppActive) {
-    handleLedOff();
-  } else {
-    handleLedOn();
-  }
-}
-
+// handle root (return)
 void handleRoot() {
   server.sendHeader("Access-Control-Allow-Origin", "*");
-  String message = "ESP32 Core is running!\n";
-  message += "User App Status: ";
-  message += isUserAppActive ? "ACTIVE" : "INACTIVE";
+  String message = "ESP32 is running!\n";
   server.send(200, "text/plain", message);
 }
 
-// --- OTA Logic (Preserved) ---
-
+// Execute OTA update from a URL
 String executeOTAFromURL(String url) {
   HTTPClient http;
   Serial.println("Starting OTA from URL: " + url);
@@ -101,6 +68,7 @@ String executeOTAFromURL(String url) {
   return "Success";
 }
 
+// Handle OTA update request (POST /ota/update?url=...)
 void handleOTAUpdate() {
   server.sendHeader("Access-Control-Allow-Origin", "*");
 
@@ -111,6 +79,8 @@ void handleOTAUpdate() {
 
   String url = server.arg("url");
   server.send(200, "text/plain", "Starting OTA update from " + url + "...");
+
+  // Give time for response to be sent
   delay(100);
 
   String result = executeOTAFromURL(url);
@@ -125,27 +95,44 @@ void handleOTAUpdate() {
 
 void setupOTA() {
   ArduinoOTA.setHostname("esp32-tartanhacks");
-  ArduinoOTA.onStart([]() { Serial.println("Start updating"); })
+
+  ArduinoOTA
+      .onStart([]() {
+        String type;
+        if (ArduinoOTA.getCommand() == U_FLASH)
+          type = "sketch";
+        else // U_SPIFFS
+          type = "filesystem";
+
+        Serial.println("Start updating " + type);
+      })
       .onEnd([]() { Serial.println("\nEnd"); })
       .onProgress([](unsigned int progress, unsigned int total) {
         Serial.printf("Progress: %u%%\r", (progress / (total / 100)));
       })
-      .onError([](ota_error_t error) { Serial.printf("Error[%u]: ", error); });
+      .onError([](ota_error_t error) {
+        Serial.printf("Error[%u]: ", error);
+        if (error == OTA_AUTH_ERROR)
+          Serial.println("Auth Failed");
+        else if (error == OTA_BEGIN_ERROR)
+          Serial.println("Begin Failed");
+        else if (error == OTA_CONNECT_ERROR)
+          Serial.println("Connect Failed");
+        else if (error == OTA_RECEIVE_ERROR)
+          Serial.println("Receive Failed");
+        else if (error == OTA_END_ERROR)
+          Serial.println("End Failed");
+      });
 
   ArduinoOTA.begin();
 }
 
-// --- Main Setup & Loop ---
-
 void setup() {
   Serial.begin(115200);
 
-  // Call user app setup
-  userAppSetup();
-
-  // Connect to WiFi
+  // connect to wifi
   Serial.println("Connecting to WiFi...");
-  WiFi.mode(WIFI_STA);
+  WiFi.mode(WIFI_STA); // Explicitly set station mode
   WiFi.begin(ssid, password);
 
   while (WiFi.status() != WL_CONNECTED) {
@@ -157,24 +144,26 @@ void setup() {
   Serial.print("IP address: ");
   Serial.println(WiFi.localIP());
 
+  // Setup OTA
   setupOTA();
+  
+  // Initialize AI functions
+  ai_test_setup();
 
+  // establish the http routes
   server.on("/", handleRoot);
-  server.on("/led/on", handleLedOn);
-  server.on("/led/off", handleLedOff);
-  server.on("/led/toggle", handleToggle);
   server.on("/ota/update", HTTP_POST, handleOTAUpdate);
   server.on("/ota/update", HTTP_GET, handleOTAUpdate);
-  server.enableCORS(true);
-  server.begin();
 
+  // allow cors for all routes
+  server.enableCORS(true);
+
+  server.begin();
   Serial.println("HTTP server started");
 }
 
 void loop() {
   ArduinoOTA.handle();
   server.handleClient();
-
-  // Delegate loop logic to user app
-  userAppLoop();
+  ai_test_loop();
 }
